[CR] TA1536

+++ routes/api/sessionMember.js
@@ -3,6 +3,8 @@
 var MessagesUtil = require('./../../util/messages');
 var sessionMemberService = require('./../../services/sessionMember');
 var inviteService = require('./../../services/invite');
+var sessionBuilderService = require('./../../services/sessionBuilder');
+var sessionBuilderSnapshotValidationService = require('./../../services/sessionBuilderSnapshotValidation');
 
 module.exports = {
   addFacilitator: addFacilitator,
@@ -12,19 +14,30 @@ module.exports = {
 function addFacilitator(req, res, next) {
   let params = req.body;
 
-  sessionMemberService.removeByRole('facilitator', params.sessionId, res.locals.currentDomain.id).then(function() {
-    sessionMemberService.createWithTokenAndColour(params).then(function(member) {
-      inviteService.createFacilitatorInvite(params).then(function() {
-        res.send({ facilitator: member, message: MessagesUtil.routes.sessionMember.addFacilitator });
+  sessionBuilderSnapshotValidationService.isFacilitatorDataValid(params.snapshot, params.accountUserId, params.sessionId, sessionBuilderService).then(function(validationRes) {
+
+    if (validationRes.isValid) {
+      sessionMemberService.removeByRole('facilitator', params.sessionId, res.locals.currentDomain.id).then(function() {
+        sessionMemberService.createWithTokenAndColour(params).then(function(member) {
+          inviteService.createFacilitatorInvite(params).then(function() {
+            res.send({ facilitator: member, message: MessagesUtil.routes.sessionMember.addFacilitator });
+          }, function (err) {
+            res.send({error:err});
+          });
+        }, function (err) {
+          res.send({error:err});
+        });
       }, function (err) {
-        res.send({error:err});
+        res.send({error: err});
       });
-    }, function (err) {
-      res.send({error:err});
-    });
+    } else {
+      res.send({ validation: validationRes });
+    }
+
   }, function (err) {
     res.send({error: err});
   });
+  
 }
 
 function getSessionMembers(req, res, next) {
